+++ src/main/java/skadistats/clarity/processor/entities/Entities.java
@@ -200,7 +200,9 @@ public class Entities {
 
     @OnDTClassesComplete
     public void onDTClassesComplete() {
-        fieldReader = engineType.getNewFieldReader(dtClasses.getPointerCount());
+        fieldReader = engineType.getNewFieldReader(
+                dtClasses.getPointerCount(),
+                dtClasses.getEntityStateFactory().getS2FieldPathType());
     }
 
     @OnReset