+++ src/main/java/skadistats/clarity/processor/sendtables/DTClasses.java
@@ -50,4 +50,8 @@ public class DTClasses {
         return pointerCount;
     }
 
+    public EntityStateFactory getEntityStateFactory() {
+        return entityStateFactory;
+    }
+
 }